Resolve randytarampi/me.words#3.
Turns out that this was a regression introduced in randytarampi/me.words#710e52e68a6f94e6e534ae3a71637f8b5a318fc2, which affected randytarampi/me.photos and randytarampi/me.blog.

The key here is that we want to ensure that all `Post`s are rendered with some height, so they can be recalculated properly (because this function gets called multiple times for a reason I do not recall, and will not investigate right now.

Prefer `window.innerHeight` over `500` because it might as well be the height of the window instead of something even more arbitrary.