fix(resume): The `resume` pug template pulls in external fontawesome CSS.
So puppeteer can load the fonts. It seems to be having a bit of difficulty loading the relative fonts...

There doesn't look like there's anything super obvious in the API docs to set the CWD on render here, but I think it's okay to cheat since the idea is that `resume.randytarampi.ca` will just redirect to `www.randytarampi.ca/resume` anyways (these local templates in `resume` will only be used to generate the PDF).