fix(www): More missing fontawesome font diagnosis per ae8b63d46850ce58cda03381bccd5a4a27f23323.
Try and force gulp to `follow` symlinks per https://github.com/gulpjs/gulp/issues/1741. It seems to already do this on my machine locally, and from all the logging I've put in I can see that the symlinks are present and the content I'm looking to copy exists, so what (if anything) is gulp doing differently on the CI boxes than on my machine?