fix(travis): Just cache nothing and always run `postinstall`.
From https://travis-ci.org/randytarampi/me/jobs/418360122#L561 it seems that `npm` is doing a lot of magic creating links and whatnot, so just always `postinstall`. At least to get these fixes out.

This change means that we'll *always* be building `lwip` on the Travis boxes. I've forgotten how slow that is, so we'll find out.