chore(www): Wean ourselves off needing the `getPhotos` and `getWords` handlers in `posts`.
I mean, they're free to run and deploy, but removing them is the easiest way to pursue #12 in `posts` right now.

Gonna push this directly to master, wait for the deploys, then remove the handlers. Anyone with a long lived cached copy is probably SOL, but really how many people could that be on this long weekend?