fix(views): Don't load resources already loaded by the stylesheets twice.
I can't quite remember why puppeteer needs to load fontawesome (specifically, the web fonts) independently of my local styles, but maybe it has something to do with how relative `url()`s in stylesheets are resolved? It clearly finds the stylesheet just fine, but it can't seem to load any of the files it references.