fix(letter): Give the footer a bit of room to breathe on `letter`s.
To mask the symptoms caused by the underlying issue in #107. I actually like the platform font variance and am not keen on making it Roboto on print, so just cheat here as I'll be rolling forward with copies generated on my laptop (with the `-apple-system` font).

Ideally we'd plant this at the bottom of a `page`, but I don't want to have to figure out how to set a page size. Better to let this be responsive for the 3 people so far that have tried to (successfully) print off my CV on varying paper sizes.