feat(resume): Redux `resume` per #69.
Give us the ability to load different resumes from some remote source – in this case right out of GitHub as a proof of concept for @canucksin and his canucksin/pantry-prototype.

The way that `resume` and `www` build their global selectors and combine their reducers is pretty gross, but I can live with it for the moment. This entire repo is pretty tangled, but there's more of a win for `www` in extending the selectors and reducers from `jsx` and `resume` than there is in dogmatically creating separate but identical containers and action creators (or otherwise decorating them and injecting consumer specific selectors). Don't even talk to me about how I'm reconciling reducers – I'm just going to justify it with something like "all these modules are ok to be opinionated and share the same state because they're in the same monorepo".

One step forward on #67, but two steps back maybe? I'll have another look at this when I tackle that, but it's a decent pattern to follow for #68.