chore(resume): Remove unnecessary reference to `NODE_CONFIG_DIR`.
To actually kick a patch release of `cea4edfabd2278e1c62ac19c0e00b6b8add604c7`. Should actually find a way of doing this in #115 without having to do this. Not the first or second or third time I've had to find something to "fix" to push a release out...