feat(letter): Redux `letter` per #68.
Use the pattern from #69 to load multiple letters from some service (GitHub raw content).

Need to figure out how exactly to support `jsx` and `js` letters since the `fetchLetter` API client is currently hardcoded for `json`, but that can come in a later commit.

For now it suffices to sketch out the functionality and to think about how to generate and serve CV/letter packages per #55.