chore: Forcibly add a `robots.txt`.
Don't know what's going on here https://travis-ci.org/randytarampi/me/jobs/429726143#L9273, creating the file in node, so it should be fine on the permissions side. What's `git` trying to do such that it can't access it? (but it could in test?).

Need to run to a coffee meeting, so 🤞 this actually works.