feat(www): `$crisp` callbacks dispatch redux actions.
For tracking purposes. It'd be nice to use Crisp's user IDs, but I don't have a set of user IDs to bind them to on my end. Could just hack around this by setting the email as an identifier, but that's obviously bad news per https://help.crisp.chat/en/article/how-to-restore-chat-sessions-with-a-token-c32v4t/