fix(jsx): Smoother scrolling through `Posts`.
Render as many `Post`s as we fetch, but fetch fewer of them and less often to minimize the potential for jitter when we're recycling DOM nodes. Adjust the "default" height from `computePostHeight` based on what (text) `Post` heights actually seem to be – don't think we'd ever get to that value, but update it while we're in here.