revert: revert: perf(travis): Let's just try building things on a single core.
Hm. So bumping the timeout in cbf2f13 "fixed" things. I don't understand how it's _sometimes_ taking 28/35s and then _most of the time_ taking ~70s. Wtf?

What are we waiting on here? Where are we losing? Are we just so busy that we can't read/write the files fast enough? Nothing looks hot locally...

This reverts commit 7da1cde5c5f379c7713590dec4a83e109f6340c9.