fix(jsx): Rejig `Error` and `ErrorWrapper` so `PostsComponent`, `ResumeComponent` and `LetterComponent` can display _specific_ errors of their own.
We can still wrap our `ReduxRoot`s with a generic `ConnectedErrorWrapper` that'll handle rendering pages for unexpected 403, 404 and 500 errors.

The motivation here is that we don't want a failed fetch on the home page to take out _all_ the content on the home page, just the relevant section (rendered by `PostsComponent`).

Individual components can handle their own errors in their own way, as `PostsComponent` does with `ENOPOSTS`, `ResumeComponent` does with `ENORESUME` and `LetterComponent` does with `ENOLETTER`.