chore(travis): Try and see if travis workspaces can speed up builds. Third time lucky?
These `paths` need to be explicitly defined.

I mean, I guess it was a bit much to assume that they'd just take the entire project directory as a workspace eh?

Don't include `node_modules` here as we're still caching those. We'll end up paying a reinstallation only on the install job where the dependencies change though, so that's cool.